requirements.txt
-e git+https://github.com/jghughesai/ai_image_analysis.git@aec49d947a9246564fed4c9b4c805af439a7924a#egg=ai_image_analysis
annotated-types==0.7.0
anthropic==0.29.0
anyio==4.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
filelock==3.15.4
fsspec==2024.6.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
jiter==0.4.2
numpy==2.0.0
openai==1.35.3
openpyxl==3.1.4
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pydantic==2.7.4
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
